Implement basic quit feature
First closes the active window.  If it is the last of the current tab,
then it closes that, and if that is the last of the tabs, then it sends
the quit message.  This doesn't handle unmodified buffers.

The sample apps now intercept the quit and close themselves.  They also
have crude parsing of the input file, so I can actually start to use the
editor as a quick viewer/replacement when I need to look at a file.