Removed cursorAfter parameters
Managing where the cursor should be after commands should be simpler
than it is.  I'm working on that here; removing cursorAfter parameters
from everywhere, and hopefully improving consistency and simplicity.
There are several tests that fail on this development branch now, but
I'm working on getting everything up and running correctly, with
comensurate improvements in the way selections/cuts/pastes work.