Improve the delay loop while waiting for keyboard
Fix for issue #15 (?)
Instead of sleeping, go back to the old method of having SDL wait for
input for 10ms - this should mean the loop executes as soon as there is
a key press.

The RefreshRequired is only checked if the timeout occurs, to test if
the editor needs to update anything.