Add a tooltip option to the marker range.
Markers can now specify if they want the indicator in the left border,
the underline under the range, and if they want the description in a
tooltip.  The mouse hovering over the range will popup the tooltip.

This is a partial solution to Issue #21, but I would also like to
broadcast a tooltip message for clients that want the option to popup a
tooltip for a hover position