Add a notify for Tooltips
Issue #21
Added a tooltip broadcast.  Clients can intercept the notification and
return a tooltip if necessary.  The notification contains the buffer and
the location where the mouse is.  From here you can determine the
tooltip.  We could extend the syntax highlighting to return the info for
the location, or even remember the keyword index for each character,
etc.  Currently the demo just pops up a tooltip for keywords &
identifiers.