Remove SDL.h from editor_imgui
Issue #16, #23

The usb_hid header file contains the USB standard scancodes, and I was
already using some of them, so just switched the 3 occurances of
SDL_SCANCODES out in the same way.

This assumes that whatever front end you are using (glfw, for example),
is sending the same USB keycodes.

SDL is no longer required to build the imgui version of the library (but
it is needed for the demo, obviously).