Fix a different ASCII character bug
The previous fix to cast char to uint caused exceptions on debug builds
on windows. Added an function to ensure that the character passed to
these functions is indeed ASCII.
Also turned on the debug build on AppVeyor to ensure that a debug build is
tested somewhere.   Travis can exit if the build takes too long, so that
is off for now.