Fixes for Zep due to integration to ImGui
I found that non-SDL didn't handle some keyboard commands correctly.
Now check for SDL and use their SCANCODES or use Win32 VK_ as the demo
does.  There may be a better way to handle all this in editor_imgui;
probably by tracking key down/up and responding to that instead.

But this seems to work in all cases.

I also found that the syntax thread wasn't waiting for completion
causing a flicker on fast running demos.  The syntax code could be more
efficient in what it changes, but that's a job for another day.