FileSystem
The file system is an abstraction over the current device's folders.  It
has simple commands for searching and read/write of files.  This is
useful if you are on a platform that doesn't have a traditional file
system, or if you want to implement a version of zep that has a
restricted set of files it can load (such as in an in-game editor, where
you only want the editor to touch files in the asset tree, or load just
files that really only exist in memory, etc.).  Another use case might
be a file editor that can only see files inside a zip file, etc.
See the methods in filesystem.h for examples.

The default filesystem is built on top of the CPP <filesystem>, as it
was before; so nothing needs to change to keep the old behavior.

fs::path is now also gone, and Zep has a simple 'ZepPath' helper class
which represents a path to a file.

The FuzzySearch method replaces the tinydir library too; removing one
more dependency.  It will eventually allow fuzzy search of a directory
tree to support CTRL+P style Vim functionality.