Extract the FileWatcher from Zep
The file watcher has platform specific code which limits the ability to
compile zep with it on.A
Instead of keeping it working inside zep, I've moved it outside to the
demo app.  The Zep Editor now has an 'OnFileChanged' method which a
client can call if it wants Zep to know about file modifications.