Issue #33 - Splits not always correct
Fixed an issue where window splits were converting a vertical split to a
horizontal one in some circumstances.  Also fixed splitting a window in
the middle of a split, so that the new window appeared along side the
existing one, instead of at the end of the stack.

Also added a bunch of comments, since when this code worked first time,
I didn't spend enough time commenting it.  It took me a while to figure
out what it was doing!

Also in this checkin, and update to hopefully fix Coverity, which is
overdue for a run.