Fix for latest windows.
DPI Scaling on the demos was broken, so I've made the demo app explicity
detect and handle the DPI of the system.  To do this, on windows it sets
an app manifest in the CMakeLists.  It also scales the font size and
sets the display scale in the editor.  This results in the editor using
the full resolution of the display; but it's just for the demo; game
engines may do something different, such as setting up a 1/2 size
framebuffer, but that's not a concern since this is just for the demo
app.

Additionally in this fix, NotModifiable is now Locked, since that better
describes a window such as the CtrlP win.

Additionally this contains a fix for closing split windows.  I think
there is still a latent bug here, but I'll take another pass at it soon.

I'll try to split up my CLs a bit, but these changes went into an app
I'm working on (which was the whole point of starting Zep in the first
place).