Bug Fix
For reference, see:
https://github.com/cmaughan/zep/pull/38

The above PR is a fix for a crash in insert mode while pressing the
delete key at the beginning of the buffer.  While the fix works, I was
concerned that this was 'too high up' in the key handling, and other
commands that do deletions might hit the same thing.  So I've fixed the
point of error; which should defend against similar scenarios.