Extending/Fixing Marker support
Markers with Background flag now correctly highlight the background of
text (useful for error reporting).  Syntax coloring has option for
background color too.

Added an option for the spacing of lines ("ZTestLineSpace 6" to try it
from Vim Command Mode).  I want to use this to give space for
underlining, etc. but it should also be a good precursor to allowing
certain lines to jump down in the buffer to make way for widget
rendering.