Modern CMake + more standard library setup
Work on improving the CMake files, and isolating the public from the
private headers.  Clients just need to include the zep/include folder
when linking to the library, and specify #include "zep/foo.h" when
including files.  This is the modern way and will stop the Zep 3rdparty
and source files polluting the namespace of the client app.

Also a fix for Release build of Qt.

I may need to figure out the single header build after this!  It should
work though