Add support for replace command
I've been missing replace for quick fixes. A perfect job or a sleepy
Sunday morning ;)  Next on my list is probably 'ctx'

The following vim commands now work:
rx (replace char under cursor with 'x')
3rx (replace next 3 chars with 'x')
vllrx (replace visual mode selected 2 characters, with x)

To support this I added Replace as a buffer command, and an undoable
command to do replacements - meaning that there are now 3 undoable
commands to interract with the underlying buffer - 'Insert', 'Delete' and 'Replace'

5 unit tests confirm the behavior.
Something that occured to me is that I need a mirror for each command that undoes it; instead of just seperate undo tests.

Also:
- Removed unnecessary imgui.ini from git