Improved CTRL+P Search
Added a project specific '.zep' hidden folder, which can optionally be
checked for ignore and include search paths.  This makes searching
faster for big projects, if you provide a narrower list of folders.

Eventually I'd like to use this feature for other project-specific
configuration, and potentially for a source database for tag lookup,
etc.