Fix README, add initial code for embedded widgets
The README is mostly complete with the new build instructions; I just
need to add a couple more Linux examples.

I also have an embeded-on-line widget example up and running; it is a
work in progress, but isolated enough not to cause any problems.  This
is for widgets that appear embedded in the text for user input or
information in a live coding scenario.