Clipboard support
Clipboard is OS/Target specific.  So the editor just broadcasts
clipboard messages when it wants to know the contents of the clipboard.
It's up to the host app to respond to the messages and populate the
clipboard or read it.

The ImGui demo here uses the 'clip' cross platform library to do the
work. I should add the same thing to the Qt demo, using whatever Qt does
for clipboard support...