Improvements to startup state + better fonts
I'm still messing a bit with how Zep starts and what the 'default' state
is.  This brings things to match vim's behavior:

- There is always at least a single tab (possibly hidden) and a single
window.
- By default there is a single [No Name] buffer, which is of type
::Default
- Hidden [No Name] buffers that are not modified are automatically
discarded.
- Adding a window to a tab containing only the default buffer will
replace it.

The focus of this change is to ensure that Zep is always in a 'valid'
state, and you can always get a tab window/window/buffer combination in
the minimal mode.

The demo now uses FreeType font engine for better text rendering in
ImGui mode.

Screenshots have been updated.