"/ ?" Searching and Range Marker improvements
In Vim mode it is now possible to search by typing / or ? and some text.
As in Vim, this will highlight all matches and show the nearest one to
the cursor.  n or N will search for next or previous after you've
pressed enter.  The usual stuff.

The RangeMarkers could probably be a bit more efficient, but I've tidied
them up to be much better than before; it's now possible to find markers
within a buffer location range using lower/upper bound searches.  I may
improve them in future too.  It's a little overkill to allocate a marker
for every range; they shold probably be in a pool or something; but for
now it works well.

Another subtle change here: when you enter Ex mode, the cursor no longer
shows/blinks, and you see line numbers instead of relative line numbers