Orca & Tree
These changes are preliminary groundwork for adding an Orca mode and a
tree.  I need the tree for (eventually) a NerdTree-like file browser,
and for (imminently) a tree based way of editing a shader setup.
The Orca mode will eventually be in the live coding project; but it is
just an experiment for now.