Beginning a Tree Mode
The tree mode is used to navigate a tree, such as a folder list, or any
kind of abstract tree hierarchy.  Eventually this will be used to
implement a NerdTree-like browser as well as a scene pass manager for a
live coding environment.

This just establishes the basic pieces and creates a default tree with
:tree.  More to come when I get time.