Keyboard Mapper & Mode improvements
Keyboard mapper

In order for different modes to handle keys differently, and eventually
to allow the user to tweak key shortcuts, I need a keyboard mapping.
A keystroke combination is stored in a tree which is analyzed as the user
types commands (in this way it can detect 'unfinished' commands, and handle combinations that should happen after
a given time if no more keys are pressed, etc.)

Refactoring modes

This is work in progress to centralize the command handling in the base
mode, and enable child modes to override commands.  Hopefully this will
make Vim and Standard modes more full featured and consistent, and
cleanup some of the input handling.