Tweaks to Syntax Coloring and add a Flash
I've added the ability to return a custom color from the syntax
highlighting and a way to ask for the syntax to flash a region of the
text.  I need this for evaluation of code fragments or the whole text,
for use during interactive repl or live coding

When this flash is engaged, the editor is refreshed at full speed until
the flash completes.  A flash range and a duration is provided.