Marker placement work
I'm working on improving the management of markers when text is
added/removed or undo / redo.  This is preliminary work; the next bit is
undo/redo.
I'm working towards a layer system (which I'l work on in a seperate
branch), but first I want to make sure that markers are correctly
located always, since they are the foundation of the layer support.