Remove Single Header from the Readme
Maintaining the single header build has become a chore, so I am
deprecating it for now.  It is easy to install Zep as a package, or just
copy/paste the source files to projects.  I don't have enough time
currently to keep the single header alive, sorry ;)

I won't actually remove the ability, just not advertise it for now.