Add FontType to ZepDisplay
This is preparation for using different font sizes in Zep.  I want the
ability to resize the main text without the rest of the UI (which is
currently tied into the global font scale in ImGui).  This will make it
easier to do a demos/performances where the viewer can see the text, and
eventually will enable markdown style display.

The next step is to create the different fonts in the Display layers and
manage their scaling independently.