Fonts & Markers
- Seperated out the Display interface and added a Font interface.  This
is for future improvements to enable different size font display
- It is now possible to provide different font sizes for predefined
types, and the main text buffer can be scaled independent of the rest of
the UI.
- Beginning markdown support: #Heading will show bigger font
- Markers now handle their own updates when the buffer changes
- Markers are outside the undo system, but flexible enough now for users
to update them as needed.  I am still working on improving marker
abilities and making them flexible enough for various scenarios.
- Integrated a simple signal/slot mechanism to make messaging easier.  I
hope this will be better than the messaging system in the long term, and
more flexible.
- Fix the bug where you couldn't 'j' down in certain wrapped text buffer
cases