Fix confusing behaviour when selecting all of empty document
When selecting all of an empty document, nothing was selected, but the
mode changed to visual mode.
This was confusing, because it's not the usual behaviour and it's not
easy to tell what mode you are in standard mode.
Now when nothing is selected, mode switches back to default (insert) mode.