Convenience Methods / Startup tweak
It is now possible to start zep without a default tab.  This makes it
easier for clients to construct things without having to worry about
that pesky 'default' buffer that appears.

In Zep the rule is that a TabWindow must have at least one Window, or it
is removed.  A Window always has a buffer or it is removed.
Now you can have a started zep without the first tab window.  It isn't
much use till you add one, but it helps with live coding
scenarios/removing all windows and resetting, etc.

Also added convenience methods for finding tabs by file path, finding
the currently active window and buffer, etc.