Tab name fix and local file save fix
The tab name didn't always correctly show the buffer name that was
active after a save
Using :w to write a new filename didn't use the file's local directory
as the relative base path.

In general, correct management of working directory and what is relative
to what needs more thought.  I think a 'Sanitize' function on the
ZepPath will help, along with some helpers to ensure that a path is
always sensible, consistent and the behavior of saving is similar enough
to Vim to make no difference.

A set of unit tests for the File path/saving stuff would be useful too.
This is tricky stuff.