Improved ImGui input key handling, and wheel
fezjo:fezjo/all PR #107
docs: add fezjo as a contributor for code

This is a more significant contribution, but after playing with it for
some time I can't find anything wrong with it, and the code is cleaner.
Mapping from ImGui is tricky, and there are lots of special case
characters.  I will keep an eye on this and see if there is anything
missing.