Tabs, Minimal Mode updates
In minimal mode, the empty line beneath a code line is also not faded;
making it easier to read sections of text.  More of the editor fades out
now.

The tabs aren't quite right, but this fix stops them flickering and
improves their display.  More to come on that one.